refactor(skills): fold enrichment into SystemSkillSyncService
Move skill keyword enrichment logic from separate
SkillTriggerEnrichmentService into SystemSkillSyncService. The sync
service already discovers skills and triggers on feed updates — having
enrichment as a separate hosted service created DI timing issues where
the IChatClientProvider wasn't reliably available during startup.

Also adds scripts/swap-daemon.sh for local daemon binary testing.
