+++ src/Netclaw.Daemon/Services/SystemSkillSyncService.cs
@@ -1,15 +1,19 @@
 using System.Security.Cryptography;
+using System.Text;
 using System.Text.Json;
+using Microsoft.Extensions.AI;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Netclaw.Actors.Skills;
+using Netclaw.Actors.Text;
 using Netclaw.Configuration;
 using Netclaw.Configuration.Feeds;
 
 namespace Netclaw.Daemon.Services;
 
 /// <summary>
-/// Syncs system skills from the feed CDN at daemon startup.
+/// Syncs system skills from the feed CDN at daemon startup, then enriches
+/// registered skills with keyword indexes for deterministic auto-loading.
 /// Runs after <see cref="Program.CopyBuiltInSkills"/> seeds offline defaults.
 /// Never blocks startup on network — if the feed is unreachable, the daemon
 /// starts with whatever skills are already on disk.
@@ -24,7 +28,9 @@ internal sealed class SystemSkillSyncService : IHostedService
     private readonly TimeProvider _timeProvider;
     private readonly ILogger<SystemSkillSyncService> _logger;
     private readonly string _daemonVersion;
-    private readonly SkillTriggerEnrichmentService? _enrichmentService;
+    private readonly IChatClientProvider? _chatClientProvider;
+
+    private static readonly TimeSpan EnrichmentTimeout = TimeSpan.FromSeconds(30);
 
     public SystemSkillSyncService(
         HttpClient httpClient,
@@ -34,8 +40,9 @@ internal sealed class SystemSkillSyncService : IHostedService
         SkillIndexContextLayer skillIndexLayer,
         TimeProvider timeProvider,
         ILogger<SystemSkillSyncService> logger,
-        SkillTriggerEnrichmentService? enrichmentService = null)
-        : this(httpClient, paths, skillSyncConfig, skillRegistry, skillIndexLayer, timeProvider, logger, BuildInfo.Version, enrichmentService)
+        IChatClientProvider? chatClientProvider = null)
+        : this(httpClient, paths, skillSyncConfig, skillRegistry, skillIndexLayer,
+            timeProvider, logger, BuildInfo.Version, chatClientProvider)
     {
     }
 
@@ -49,7 +56,7 @@ internal sealed class SystemSkillSyncService : IHostedService
         TimeProvider timeProvider,
         ILogger<SystemSkillSyncService> logger,
         string daemonVersion,
-        SkillTriggerEnrichmentService? enrichmentService = null)
+        IChatClientProvider? chatClientProvider = null)
     {
         _httpClient = httpClient;
         _paths = paths;
@@ -59,7 +66,7 @@ internal sealed class SystemSkillSyncService : IHostedService
         _timeProvider = timeProvider;
         _logger = logger;
         _daemonVersion = daemonVersion;
-        _enrichmentService = enrichmentService;
+        _chatClientProvider = chatClientProvider;
     }
 
     public async Task StartAsync(CancellationToken cancellationToken)
@@ -292,6 +299,7 @@ internal sealed class SystemSkillSyncService : IHostedService
     /// <summary>
     /// Re-scans the entire skills directory and rebuilds the registry + context layer.
     /// Called after sync completes (or fails) to ensure the agent sees current skills.
+    /// Then enriches all skills with keyword indexes for deterministic auto-loading.
     /// </summary>
     private void RescanAndUpdateIndex()
     {
@@ -302,14 +310,199 @@ internal sealed class SystemSkillSyncService : IHostedService
         _skillIndexLayer.Update(_skillRegistry.GenerateCompressedIndex());
         _logger.LogInformation("Skill index updated ({SkillCount} skills)", _skillRegistry.GetAll().Count);
 
-        // Trigger keyword enrichment for auto-loading (fire and forget)
-        if (_enrichmentService is not null)
-            _ = _enrichmentService.EnrichAllAsync();
+        // Enrich keywords for deterministic skill auto-loading (fire and forget)
+        _ = Task.Run(async () =>
+        {
+            try
+            {
+                await EnrichAllSkillsAsync();
+            }
+            catch (Exception ex)
+            {
+                _logger.LogWarning(ex, "Skill keyword enrichment failed");
+            }
+        });
+    }
+
+    // ── Skill keyword enrichment ──
+
+    private async Task EnrichAllSkillsAsync()
+    {
+        var skills = _skillRegistry.GetAll();
+        if (skills.Count == 0)
+            return;
+
+        _logger.LogInformation("Starting skill keyword enrichment ({Count} skills)", skills.Count);
+        Directory.CreateDirectory(_paths.SkillKeywordCacheDirectory);
+
+        foreach (var skill in skills)
+        {
+            try
+            {
+                await EnrichSkillAsync(skill);
+            }
+            catch (Exception ex)
+            {
+                _logger.LogWarning(ex, "Enrichment failed for {SkillName}, using fallback", skill.Name);
+                ApplyFallbackKeywords(skill);
+            }
+        }
+
+        _logger.LogInformation("Skill enrichment complete: {EnrichedCount}/{TotalCount} skills enriched",
+            _skillRegistry.GetEnrichedKeywords().Count, skills.Count);
+    }
+
+    private async Task EnrichSkillAsync(SkillEntry skill)
+    {
+        if (!File.Exists(skill.FilePath))
+        {
+            _logger.LogWarning("Skill file not found: {FilePath}", skill.FilePath);
+            ApplyFallbackKeywords(skill);
+            return;
+        }
+
+        var content = await File.ReadAllTextAsync(skill.FilePath);
+        var contentHash = ComputeSha256(content);
+
+        // Check cache
+        var cached = LoadCachedKeywords(skill.Name, skill.Version, contentHash);
+        if (cached is not null)
+        {
+            _skillRegistry.SetEnrichedKeywords(skill.Name, cached);
+            _logger.LogDebug("Loaded cached keywords for {SkillName} ({Count} keywords)",
+                skill.Name, cached.Count);
+            return;
+        }
+
+        // Try sidecar LLM
+        var keywords = await GenerateKeywordsAsync(skill.Name, content);
+        if (keywords is null || keywords.Count == 0)
+        {
+            _logger.LogWarning("Sidecar returned no keywords for {SkillName}, using fallback", skill.Name);
+            ApplyFallbackKeywords(skill);
+            return;
+        }
+
+        _skillRegistry.SetEnrichedKeywords(skill.Name, keywords);
+        SaveCachedKeywords(skill.Name, skill.Version, contentHash, keywords);
+        _logger.LogInformation("Enriched {SkillName} with {Count} keywords via LLM",
+            skill.Name, keywords.Count);
+    }
+
+    private async Task<HashSet<string>?> GenerateKeywordsAsync(string skillName, string content)
+    {
+        if (_chatClientProvider is null)
+            return null;
+
+        try
+        {
+            var client = _chatClientProvider.GetClient(ModelRole.Main);
+            using var cts = new CancellationTokenSource(EnrichmentTimeout);
+
+            var messages = new List<ChatMessage>
+            {
+                new(ChatRole.System, """
+                    You are a keyword extraction assistant. Given a skill document that
+                    describes when and how an AI agent should behave, generate a comprehensive
+                    list of single words and short phrases (2-3 words max) that a USER might
+                    say in a message when they would benefit from this skill's guidance.
+
+                    Include:
+                    - Action verbs users would use (buy, find, compare, book, etc.)
+                    - Domain nouns the skill covers (price, product, flight, restaurant, etc.)
+                    - Common user phrasings (how much, best deal, where to, etc.)
+                    - Adjectives users might use (cheap, expensive, best, good, etc.)
+
+                    Return ONLY the keywords, one per line, lowercase. No numbering, no
+                    explanations, no categories. Just the words.
+                    """),
+                new(ChatRole.User, content)
+            };
+
+            var response = await client.GetResponseAsync(messages, cancellationToken: cts.Token);
+            var text = response.Messages[^1].Text ?? string.Empty;
+
+            if (string.IsNullOrWhiteSpace(text))
+                return null;
+
+            var keywords = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
+            foreach (var line in text.Split('\n', StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries))
+            {
+                foreach (var token in TextTokenizer.Tokenize(line))
+                    keywords.Add(token);
+            }
+
+            return keywords.Count > 0 ? keywords : null;
+        }
+        catch (OperationCanceledException)
+        {
+            _logger.LogWarning("Enrichment LLM timed out for {SkillName}", skillName);
+            return null;
+        }
+        catch (Exception ex)
+        {
+            _logger.LogWarning(ex, "Enrichment LLM call failed for {SkillName}", skillName);
+            return null;
+        }
+    }
+
+    private void ApplyFallbackKeywords(SkillEntry skill)
+    {
+        var text = (skill.Triggers ?? string.Empty) + " " + skill.Description;
+        var keywords = TextTokenizer.Tokenize(text).ToHashSet(StringComparer.OrdinalIgnoreCase);
+        if (keywords.Count > 0)
+            _skillRegistry.SetEnrichedKeywords(skill.Name, keywords);
+    }
+
+    // ── Keyword cache I/O ──
+
+    private record KeywordCacheEntry(string SkillName, string? Version, string ContentHash, string[] Keywords);
+
+    private string GetKeywordCachePath(string skillName, string? version)
+    {
+        var fileName = version is not null ? $"{skillName}-{version}.json" : $"{skillName}.json";
+        return Path.Combine(_paths.SkillKeywordCacheDirectory, fileName);
+    }
+
+    private HashSet<string>? LoadCachedKeywords(string skillName, string? version, string contentHash)
+    {
+        var path = GetKeywordCachePath(skillName, version);
+        if (!File.Exists(path))
+            return null;
+
+        try
+        {
+            var json = File.ReadAllText(path);
+            var entry = JsonSerializer.Deserialize<KeywordCacheEntry>(json);
+            if (entry is null || entry.ContentHash != contentHash)
+                return null;
+
+            return new HashSet<string>(entry.Keywords, StringComparer.OrdinalIgnoreCase);
+        }
+        catch (Exception ex)
+        {
+            _logger.LogWarning(ex, "Failed to load keyword cache for {SkillName}, treating as miss", skillName);
+            return null;
+        }
+    }
+
+    private void SaveCachedKeywords(string skillName, string? version, string contentHash, HashSet<string> keywords)
+    {
+        try
+        {
+            var entry = new KeywordCacheEntry(skillName, version, contentHash, keywords.ToArray());
+            var json = JsonSerializer.Serialize(entry, new JsonSerializerOptions { WriteIndented = true });
+            File.WriteAllText(GetKeywordCachePath(skillName, version), json);
+        }
+        catch (Exception ex)
+        {
+            _logger.LogWarning(ex, "Failed to save keyword cache for {SkillName}", skillName);
+        }
     }
 
     internal static string ComputeSha256(string content)
     {
-        var bytes = System.Text.Encoding.UTF8.GetBytes(content);
+        var bytes = Encoding.UTF8.GetBytes(content);
         var hash = SHA256.HashData(bytes);
         return Convert.ToHexStringLower(hash);
     }
